Basic Layout Creation
This commit contains the layout for the Mercury level that holds the
current following. The imageview collected from the shipoutline.png. In
addition now has basic movements across the screen, however there is a
possibility that the ship will exit off screen and never come
back...needs work on boundries